requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
allure-pytest==2.13.3
allure-python-commons==2.13.3
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
constantly==15.1.0
coverage==7.4.1
cryptography==41.0.4
cssselect==1.2.0
dnspython==2.4.2
filelock==3.12.4
frozenlist==1.4.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
multidict==6.0.4
openai==0.28.1
outcome==1.3.0.post0
packaging==23.2
parsel==1.8.1
pluggy==1.3.0
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pymongo==4.5.0
pyOpenSSL==23.2.0
PySocks==1.7.1
pytest==7.4.4
pytest-cov==4.1.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.11.0
selenium==4.17.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tldextract==5.0.0
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
Twisted==22.10.0
typing_extensions==4.9.0
urllib3==2.0.6
w3lib==2.1.2
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.1
